Move tests to unit directory
