+++ packages/salesforcedx-visualforce-language-server/package.json
@@ -30,7 +30,7 @@
     "source-map-support": "^0.4.15"
   },
   "scripts": {
-    "compile": "tsc -p ./ && shx cp -R test/fixtures out/test",
+    "compile": "tsc -p ./ && shx cp -R test/unit/fixtures out/test/unit",
     "lint": "tslint --project .",
     "watch": "tsc -watch -p .",
     "clean":
@@ -38,7 +38,7 @@
     "test":
       "./node_modules/.bin/nyc ./node_modules/.bin/_mocha --recursive out/test --reporter mocha-multi-reporters --reporter-options configFile=../../config/mochaConfig.json",
     "test:unit":
-      "./node_modules/.bin/nyc ./node_modules/.bin/_mocha --recursive out/test --reporter mocha-multi-reporters --reporter-options configFile=../../config/mochaConfig.json"
+      "./node_modules/.bin/nyc ./node_modules/.bin/_mocha --recursive out/test/unit --reporter mocha-multi-reporters --reporter-options configFile=../../config/mochaConfig.json"
   },
   "nyc": {
     "reporter": ["text-summary", "lcov"]
